Fix WP < 5.5 compatibility due to incompatible method signatures.
WordPress 5.5 added a new parameter to `WP_Upgrader::download_package()`. WP-CLI
subclasses the `Core_Upgrader` class to customize this behavior. This causes a
PHP warning on 7.4 and fatal error on 8.0 if the method signatures do not match.

In #166 the method signature was adjusted to account for this new parameter,
however this causes the same incompatible signature issue on sites running WP
versions before 5.5.

This commit introduces a trait-based compatibility shim to use the correct method
signature depending on the WP version being run. This is patterned off the
implementation used in the `FeedbackMethodTrait`.